requirements.txt
asgiref==3.8.1
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.1
coreapi==2.3.3
coreschema==0.0.4
Django==5.1.4
django-activity-stream==2.0.0
django-cors-headers==4.7.0
django-environ==0.11.2
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Markdown==3.7
MarkupSafe==3.0.2
psycopg2==2.9.10
PyJWT==2.10.1
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.22.3
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
